Revert "Fix Json error with deserilaizer. Add icons to results page"
This reverts commit f3431ed3ff708e31cfb5eeb64c9f5b54f1725e3d.

+++ Recipe.Web/Shared/Models/SpoonacularRecipe.cs
@@ -282,7 +282,7 @@ namespace Recipe.Web.Data.Models
         public double AverageRating { get; set; }
 
         [JsonPropertyName("ratingCount")]
-        public double RatingCount { get; set; }
+        public long RatingCount { get; set; }
 
         [JsonPropertyName("score")]
         public double Score { get; set; }
